fix: set fip control plane load balancer ip mode to proxy
With IP mode VIP (the default, and what is used right now),
`kube-proxy` forwards the traffic straight to the service endpoints,
so health checks performed through the FIP don't actually hit the required node.
With IP mode to `proxy`, traffic gets sent directly to the FIP,
and therefore the node.